+++ banglaspeech2text/utils/download_path.py
@@ -1,6 +1,5 @@
 # Determine Download Path
 import os
-import sys
 
 
 def get_user_path():
@@ -9,14 +8,11 @@ def get_user_path():
         if os.environ['BanglaSpeech2Text']:
             return os.environ['BanglaSpeech2Text']
 
-    if sys.platform == 'win32':
-        return os.path.expanduser('~')
-    else:
-        return os.path.expanduser('~/Downloads')
+    return os.path.expanduser("~")
 
 
 def get_app_path(app_name: str) -> str:
     path = os.path.join(get_user_path(), app_name)
     if not os.path.exists(path):
-        os.mkdir(path)
+        os.makedirs(path)
     return path