+++ example.py
@@ -1,9 +1,11 @@
 from banglaspeech2text import available_models, Model
 
 
-models = available_models()
+models = available_models(True)
+print(models)
+# file_path = r"F:\Code\Python\rodela_bot\app\data\উদ্দেশ্য_licence.wav"
 
-model = Model(models[0], verbose=True)
-model.load()
+# model = Model(models[0], verbose=True)
+# model.load()
 
-print(model)
+# print(model(file_path))